services/ingest-odds/pyproject.toml
[tool.poetry]
name = "ingest-odds"
version = "0.1.0"
description = "Live odds streamer for multiple bookmakers"
authors = ["faztyG <verselabproject@gmail.com>"]
readme = "README.md"
packages = [{ include = "ingest_odds" }]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
aiohttp = "^3.9.5"
nats-py = "^2.11.0"
redis = "^6.4.0"
orjson = "^3.11.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
